backend/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langgraph-checkpoint-postgres

# Database and Persistence
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API and Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# WebSocket for real-time updates
websockets>=12.0

# CORS and middleware
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0

# MCP Server (for mcp-server directory)
mcp>=0.9.0

# Utilities
httpx>=0.27.0
tenacity>=8.2.0
